Refactored postgres proxy to support TLS
