+++ backend/src/test/kotlin/dev/kviklet/kviklet/proxy/TemplatePostgresProxyTest.kt
@@ -1,0 +1,49 @@
+package dev.kviklet.kviklet.proxy
+
+import dev.kviklet.kviklet.db.EventAdapter
+import dev.kviklet.kviklet.db.ExecutionRequestAdapter
+import org.junit.jupiter.api.AfterEach
+import org.junit.jupiter.api.BeforeEach
+import org.springframework.boot.test.context.SpringBootTest
+import org.springframework.test.context.ActiveProfiles
+import org.testcontainers.containers.PostgreSQLContainer
+import dev.kviklet.kviklet.proxy.helpers.*
+import org.junit.jupiter.api.Test
+import org.springframework.beans.factory.annotation.Autowired
+import java.sql.Connection
+
+@SpringBootTest
+@ActiveProfiles("test")
+class TemplatePostgresProxyTest {
+    @Autowired
+    lateinit var executionRequestAdapter: ExecutionRequestAdapter
+    @Autowired
+    lateinit var eventAdapter: EventAdapter
+    private lateinit var directConnection : Connection
+    private lateinit var proxy : ProxyInstance
+    private lateinit var postgresContainer : PostgreSQLContainer<Nothing>
+    @BeforeEach
+    fun setup() {
+        postgresContainer = PostgreSQLContainer<Nothing>("postgres:13").apply {
+            withDatabaseName("testdb")
+            withUsername("test")
+            withPassword("test")
+        }
+        postgresContainer.start()
+        while(!postgresContainer.isRunning) { Thread.sleep(1000) }
+        this.directConnection = directConnectionFactory(postgresContainer)
+        this.proxy = proxyServerFactory(postgresContainer, executionRequestAdapter, eventAdapter)
+    }
+    @AfterEach
+    fun tearDown() {
+        this.proxy.proxy.shutdownServer()
+        this.proxy.connection.close()
+        this.postgresContainer.stop()
+    }
+    @Test
+    fun `MockPostgresProxyTest`() {
+        assert(true)
+    }
+
+}
+
